backend/sharethis/requirements.txt
boto3==1.20.54
coloredlogs==15.0.1
flask==2.0.2
gunicorn==20.1.0
marshmallow==3.14.1
psycopg2-binary==2.9.3
requests==2.27.1
sqlalchemy==1.4.31
flask-cors
azure-storage-blob==12.10.0
